refactor: createHtml.ts 분리로 HTML 생성과 DOM 삽입 책임 분리
